+++ NEWS.md
@@ -1,5 +1,8 @@
 # teal.data 0.5.0.9001
 
+### Enhancements
+* `col_relabel` supports `NA` to remove labels (similar to the `col_labels<-`).
+
 ### Bug fixes
 * Fixed bug in `col_labels` causing incorrect label names to be returned when input data contains named label attributes.